+++ lib/op-state-get.c
@@ -88,7 +88,7 @@ int wimaxll_state_get(struct wimaxll_handle *wmx)
 			WIMAX_GNL_OP_STATE_GET, WIMAX_GNL_VERSION) == NULL) {
 		result = -ENOMEM;
 		wimaxll_msg(wmx, "E: STATE_GET: error preparing message: "
-			  "%d 0x%08x\n", result, result);
+			  "%zd 0x%08x\n", result, (unsigned int) result);
 		goto error_msg_prep;
 	}
 	nla_put_u32(msg, WIMAX_GNL_STGET_IFIDX, (__u32) wmx->ifidx);
@@ -107,6 +107,6 @@ error_msg_send:
 	nlmsg_free(msg);
 error_msg_alloc:
 error_not_any:
-	d_fnend(3, wmx, "(wmx %p) = %d\n", wmx, result);
+	d_fnend(3, wmx, "(wmx %p) = %zd\n", wmx, result);
 	return result;
 }
